% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_all}
\alias{config_all}
\title{Return the full in-memory config}
\usage{
config_all()
}
\value{
A named list containing all loaded config values.
}
\description{
Return the full in-memory config
}
\examples{
\dontrun{
load_config("config.yml")
str(config_all())
}
}
\seealso{
\code{\link[=load_config]{load_config()}}, \code{\link[=config_get]{config_get()}}
}
