% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_clear}
\alias{config_clear}
\title{Clear the in-memory config}
\usage{
config_clear()
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Clear the in-memory config
}
\examples{
\dontrun{
config_clear()
}
}
\seealso{
\code{\link[=load_config]{load_config()}}
}
