% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{is_secret}
\alias{is_secret}
\title{Test whether an object is an orchrd_secret}
\usage{
is_secret(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} is an \code{orchrd_secret}, otherwise \code{FALSE}.
}
\description{
Test whether an object is an orchrd_secret
}
\seealso{
\code{\link[=secret]{secret()}}, \code{\link[=reveal]{reveal()}}
}
